fix: testing and extraction of files larger than 2G
SVN: 11800

+++ source/java/ch/systemsx/cisd/hdf5/tools/HDF5ArchiveTools.java
@@ -1098,7 +1098,7 @@ public class HDF5ArchiveTools
             byte[] buffer)
     {
         final CRC32 crc32Digest = new CRC32();
-        int offset = 0;
+        long offset = 0;
         while (offset < size)
         {
             final int n =
@@ -1325,7 +1325,7 @@ public class HDF5ArchiveTools
         final CRC32 crc32 = new CRC32();
         try
         {
-            int offset = 0;
+            long offset = 0;
             while (offset < size)
             {
                 final int n =
